pyproject.toml
[tool.poetry]
name = "near-bs"
version = "0.1.0"
description = "identify the nearest 5G base station for each 4G base station, facilitating the establishment of an anchor point from the 4G side to ensure the seamless operation of non-standalone 5G (NSA) deployments"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "near_bs"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
oracledb = "^2.4.1"
python-dotenv = "^1.0.1"
pandas = "2.0.3"
scipy = "1.10.1"
geopy = "^2.4.1"
enm-client-scripting = {path = "../local_packages/enm_client_scripting-1.22.2-py2.py3-none-any.whl"}
paramiko = "^3.5.0"

[tool.poetry.scripts]
mobatch = "near_bs.scripts.mobatch:mobatch"
get-csv = "near_bs.scripts.csv:get_csv"

[tool.poetry.group.dev.dependencies]
black = "24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
flake8-print = "^5.0.0"
flake8-docstrings = "^1.7.0"
mypy = "^1.13.0"
pandas-stubs = "2.0.3.230814"
types-paramiko = "^3.5.0.20240928"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
